Move email functions into mail_api
